pyproject.toml
[tool.poetry]
name = "connectors"
version = "0.1.0"
description = ""
authors = ["Heiko Kulinna <Heikokulinna@gmx.de>"]

# update lockfile
# poetry update --lock -vvv

[tool.poetry.dependencies]
python = "^3.8"


virtualenv = "== 20.16.2"
pandas = "*"
numpy = "*"

adlfs = "*"
azure-kusto-data = "*"
azure-storage-blob = "*"

python-dotenv = "^0.17.0"

pyarrow = "^6.0.1"
"yaml-1.3" = ">= 0.1.0"


# importlib-metadata = {version = "*", markers = "python_vresion < \ "3.8"\""}



[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
isort = "^5.8.0"
pre-commit = "^2.12.1"
pylint = "^2.12.2"



[tool.black]
line-length = 79


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
